requirements.txt
beautifulsoup4==4.11.2
python-telegram-bot==20.3
requests==2.28.2
